feat: Added react-toastify library for toasts.
Added the react-toastify library to give a visual indicator to the user
when he updates his profile, instead of just having the submit button
disappear without any indication whatsoever. I have used the
react-toastify library here as it had a very easy to use and low code
interface, compared to react-bootstrap's Toast. It also looked a lot
better than react-bootstrap's Toast, and didn't interfere with the
existing UI of the application.